renamed integration example
